+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+## 1.0.0 2021-03-19
+
+-   Null safety
+-   Fix example to use local package
+-   Improve README
+
 ## 0.1.1 2021-01-06
 
-* Fix example name per publishing guidelines
+-   Fix example name per publishing guidelines
 
 ## 0.1.0 2021-01-06
 
-* First release
+-   First release